backend/requirements.txt
psycopg2-binary>=2.9.9
flask==3.0.0
flask-cors==4.0.0
pandas>=2.2.3
numpy>=1.26.4
scikit-learn>=1.5.0
joblib>=1.3.2
python-dotenv>=1.0.1